8-neighbor multiple flow direction routing
The multi-direction 4-neighbor algorithm used in the original version of DHSVM can cause problems when the 8-neighbor average slope from slope_aspect() conflicts with the true flow path due to local terrain heterogeneity, leading to infinite loops and unphysical surface ponding (c.f. Mountain Hydrology white paper: https://mountainhydrology.com/mh004_flow-routing-loops/). However, the D8 algorithm that was also implemented in later versions of DHSVM does not allow flow to spread out, which leads to over-concentrated and unrealistic hillslope flow paths.

This update implements a multi-direction 8-neighbor flow algorithm that routes water to all downhill cells in proportion to the local slope. Note that the effective flow width now has a circular rather than rectangular interpretation: if only one cell is downhill, the effective flow width is dx * PI/8, and if all 8 cells are downhill, the effective flow width is dx * PI. For the new algorithm to work, it is assumed that the input DEM is hydrologically corrected for the D8 case with a minimum slope, so that all cells have at least one downhill neighbor.

Note also that this commit completely removes the original 4-neighbor and D8 flow routing options. If desired, these other flow routing options would need to be added back with an option in the settings header or config file.