Infiltration out of the bottom of stream channels
This implements the first working draft of a new method for calculating infiltration out of the bottom of stream channels (i.e., re-infiltration of streamflow into the soil). Historically, there was no lateral flow allowed below stream channels, which meant that the water table was always at or above the channel cut depth (but of course this causes its own problem in arid terrain, see earlier commit enabling lateral flow below channels). Once lateral flow was implemented below channels, in some places the water table would drop significantly below the channel cut, and thus the channels were effectively impermeable aqueducts bridging water over dry terrain.

In this version, a new function is presented, channel_grid_infiltration() in channel_grid.c, that calculates infiltration out of the bottom of channels iff the water table is below the channel cut depth. This is an approximation--what if the water table is above the channel cut depth, but below the channel water surface?--but it seems to work and sufficiently redistribute water out of "losing" streams in arid terrain. This version REQUIRES a new column of data in the stream map file immediately after the channel cut width and before the (optional) channel aspect. So, the 7th column of the stream map file is now the infiltration rate for each channel/grid cell intersection (unit: mm/s, NOT m/s), which can be spatially variable (perhaps calculated from the soil maps). Determining an appropriate infiltration rate may take further work, but back-of-the-envelope tests worked well with a value of 0.012 mm/s (1 m/d) for all rows in the stream map file. Setting the infiltration rate to 0.0 should reproduce the classical behavior, but a value is required for input.

N.B.: this new functionality is completely unrelated to the similar-sounding road network infiltration! The (required) stream infiltration rate is for re-infiltration from channels with non-zero liquid water storage (i.e., current streamflow) and is read from the stream map file; the (optional) road infiltration rate is for rain infiltration on the (optional) road network and is read from the class file.