Fixed deep layer moisture initialization
InitSoilMap() only copied porosity and field capacity for the root zone layers, but InitModelState() tried to make sure the SoilMap[x][y].Moist[i] was at least equal to the field capacity for the deep layer, which was uninitialized, leading to an inconsistency. This was only really a problem if the user specifies an initial soil state with very low (i.e., zero) soil moisture, but this can be a desirable specification in arid climates so that the model automatically sets the deep layer moisture equal to field capacity at startup. This commit fixes the initialization of the porosity and field capacity maps so the deepest root zone parameters are copied into the deep layer, and also fixes InitModelState so that it compares the root zone layer moisture with the wilting point of each respective layer instead of just the deepest one.