Add back option for kinematic surface routing
This is based on the original v3.1.1 DHSVM code, which included the sediment module (c.f. https://github.com/pnnl/DHSVM-PNNL/commit/8e8cc5814da5083b2dfaad9a43debb3b287908e0). The sediment module was subsequently removed from the main branch of the DHSVM GitHub, which also removed the functionality for overland kinematic wave flow routing even without sediment. In this commit, I have added back the parts of the code necessary for the kinematic wave routing (with Overland Routing = KINEMATIC) and adapted parts of the original version for compatibility with new features such as impervious areas and detention storage. This commit does NOT restore the sediment functionality. Because of the way the soil table is read, this commit causes a breaking change with old config files because the updated version requires a parameter "Mannings N" for each soil class that must be specified as a placeholder even if Overland Routing = CONVENTIONAL.