Add stream channel and surface pond evaporation
In arid landscapes, stream channels are sometimes one of the only sources of near-surface moisture and might contribute substantially to total ET in certain areas. This commit implements a new function ChannelEvaporation() that allows evaporation from the surface of stream channels using the Penman equation based on the SoilEvaporation() function. Thus, stream channels can lose water in 3 ways: outflow to downstream segments, re-infiltration to the subsurface, or evaporation to the atmosphere. Stream evaporation is tracked in the Mass.Balance file under the name "ChannelEvap" and reported in the Mass.Final.Balance in addition to the Stream.Flow file. The Stream.Flow output fields are now: inflow, lateral_inflow, outflow, storage [Totals only], storage change, infiltration, evaporation, error [Totals only].

Additionally, this commit implements the function PondEvaporation() which removes IExcess up to the rate satisfying EPot. Both of these functions are included along with the SoilEvaporation() function in the new SurfaceEvaporation.c file, which supersedes the legacy SoilEvaporation.c file.