Refine stream-groundwater interactions
This commit fixes various issues with how I initially implemented the stream-groundwater interaction. The biggest change is that a grid cell can only receive channel infiltration from water originating at a higher elevation, i.e., water flowing into the local segment as upstream inflow or lateral inflow. Previously, the maximum infiltration was based on total segment storage, which could cause water to "teleport" uphill if some of the channel storage was contributed from lower-elevation lateral inflow.

Additionally, this commit considerably revisits how the subsurface gradient is calculated between the water table and the stream channel. In this version, the gradient is a function of channel width, so that the effective hydraulic head is calculated from midway between the channel bank and the edge of the grid cell.