Channel infiltration can include inflow
This commit represents a considerable re-thinking of the channel infiltration logic. Previously, infiltration was only available from water "stored" in a channel, but with the DHSVM channel routing scheme, the segment "storage" does not really represent the amount of water "in" a channel, but rather an effective storage related to the residence time. As a result, lots of water can flow in and out of a segment without increasing the storage. In this version, inflow from upstream segments is considered fully available for infiltration, i.e., if a channel flows into a dry, porous area, the full inflow volume can re-infiltrate. This results in much larger surface/groundwater exchanges and more realistic channel disconnection in arid environments.