On-the-fly infiltration during channel routing
This was tricky, and required implementing a cross-walk strategy between the channel network and the channel map, but it was ultimately necessary for physical fidelity and consistency. Water entering a channel as top-down inflow from other segments should be immediately available for infiltration to any grid cell in the new (lower) channel segment. Thus, the available water for infiltration and the infiltration amount itself both need to be calculated on-the-fly during stream channel routing. This requires a way of identifying all map cells that intersect each channel segment, which is now handled by the new function channel_combine_map_network.