Use anisotropy ratio to create vertical Ksat maps
The option "Vertical Ksat Source" in the OPTIONS section can now be specified as "TABLE" (default) or "ANISOTROPY." In the former case, nothing should change compared to the legacy model behavior, and the vertical conductivity of each layer is read from the soil lookup table. If the ANISOTROPY option is specified, the vertical hydraulic conductivity in each cell is derived from the mean lateral conductivity of each soil layer (calculated from the surface conductivity and the layer depth) divided by the (newly required) "Vertical Anisotropy" value from the soil lookup table. In this case, the maximum surface infiltration is also set equal to the surface layer vertical conductivity.

The goal of this option is to enable easy calibration of the vertical and lateral conductivity in a map-based fashion. Either the anisotropy ratio can be fixed to a constant, in which case both Klat and Kvert are scaled together, or the anisotropy ratio can be a new calibration parameter that is dependent on the Klat map and the exponential decrease.