Map-based exponential decrease
This was a fairly simple change to make the exponential decrease in Ksat with water table depth into a map-based parameter. Analogous to the Ksat parameter, it is still required as an input in the soil lookup table, but it is overwritten if Soil Exponential Decrease Map File = [file] in float format with single layer of exponential decrease values.