Replace assertions
Assertions are considered bad practice in production code. It is up to the user to check the model outputs (especially with the new default -Ofast optimization) and ensure there are no NANs or other implausible values. Minor rounding errors that previously might have triggered non-negativity assertions are better handled by resetting the parameter in question to zero and tracking numerical error as usual.