Improve stability of water table and streamflow
This is a complicated fix to multiple interconnected problems...

With certain combinations of parameters and model conditions (in particular low porosity, deep soil, high conductivity, and low streamflow), the water table routing option could cause water to flow back and forth between cells that have very similar topographic elevations. This was a result of assuming a quasi-steady-state gradient based on differences in water table elevation. In flat terrain with low porosity and deep soil, a small amount of SatFlow could cause the water table to fluctuate by several meters between timesteps, causing an unphysical groundwater oscillation due to reversal of the local hydraulic gradients.

This version implements several new procedures:
(1.) If the change in water table elevation between timesteps is large than the topographic or hydraulic downhill gradient, the cell water table elevation is smoothed with a recursive time-step average, i.e., WaterLevel = average(WaterLevel, WaterLevelLast).
(2.) The updated (and potentially time-averaged) WaterLevel is used for determining the routing parameters such as transmissivity instead of the instantaneous water table depth.
(3.) If the differences in water level between cells are on the order of the soil depth (so changes in water table can cause changes in flow routing direction), an additional procedure is implemented during lateral routing of SatFlow to only allow as much flow as would cause the uphill cell to come into equilibrium with the downhill cell.

Additionally, this commit includes some changes to the hyporheic exchange scheme to improve streamflow stability (from timestep to timestep and channel segment to segment). The main change is that infiltration capacity into the subsurface is only calculated based on the volume of dry soil below the channel cut, not the entire area of the grid cell. This makes the channel infiltration rate independent of the selected grid cell resolution.