Mitigate time-step dependence of infiltration capacity
Two factors determine how much water can infiltrate from channels on a given time-step: the infiltration rate (from the vertical conductivity), and the capacity of non-saturated soil directly beneath the channel. Of these two, the capacity is usually the limiting factor. Thus, with a quicker model time-step, the capacity can be reset more frequently, causing the hyporheic behavior of the model to be somewhat dependent on the (arbitrary) user-selected time-step.

This updated version considers the lateral transmissivity within the part of the grid cell where water is flowing away from the infiltration zone in order to determine the rate of diffusion from the soil column directly below the channel to the rest of the grid cell. This in defines a local water table beneath the channel (equal to or shallower than the grid cell water table) that in turn defines a dynamic infiltration capacity on each time-step.

This change only really matters if the model time-step is fairly small, probably < 1 hour. Nevertheless, it makes the overall hyporheic scheme more physical and less arbitrary in some cases.