Fix precip multiplier consistency
At this point I'm not sure what the precipitation multiplier was originally intended to do--it seems weird. In my version, it's intended as a simple (domain-average) scale factor to offset biased precipitation fields, i.e., Ptrue = Pmeteorology * Pmultiplier. I previously implemented this for the case that PRISM = TRUE, and this commit fixes it so that the Precipitation Multiplier Map works the same when PRISM = FALSE.