Add capability for 3D wind redistribution of snow
This commit implements the first working version of my new blowing snow / wind redistribution scheme for DHSVM. It is largely based on the Prairie Blowing Snow Model (PBSM) and Canadian Hydrology Model (CHM). Relevant citations are in the code and will be summarized later.

The main difference with this scheme is that it can ingest an arbitrary number of vertical wind layers, each with a distance XYZ velocity field. This allows coupling of DHSVM to a computational fluid dynamics wind model, such as the OpenFOAM implementation in WindNinja. The advantage of this approach is that blowing snow does not need to be in equilibrium with the local ground saltation concentation--rather, snow blowing off of steep terrain becomes airborne and falls at a rate determined by the settling velocity and vertical wind component. This enable spindrift or "snow plumes" to originate from mountaintops and propagate out over cirques, avoiding the problem exhibited by other models where deep snow is concentrated immediately downwind of the first terrain obstacle.

Many additional improvements and bug fixes will doubtless be necessary, so this should be considered at best a tentative "v1" of the DHSVM blowing snow module.