Pattern-based snow redistribution
This extends the functionality of the "PRISM" component of DHSVM, which was originally intended to prescribe precipitation patterns (the PRISM name for this functionality is unfortunate, because it works with any source of precip pattern data). The PRISM multiplier maps apply to ALL precipitation, both snow and rain. Snow patterns (e.g., drifts) can be simulated in DHSVM by "tricking" the precip so that the snow magically ends up in approximately the right place. This was previously accomplished by imprinting drift patterns into the cold-season PRISM maps, but this has the unintended consequence of also redistributing rain according to the drift pattern. It is preferable that rain should be distributed according to the raw PRISM pattern, and snowfall should be distributed according to some combination of the PRISM pattern (precipitation) and the drift pattern (redistribution).

This version now accepts a snow pattern map, which is used to spatially re-scale snowfall only and does not apply to rain. Note that the snow pattern is applied in addition to the PRISM (precip) pattern, so the true snowfall is (PRISM precip pattern relative to station) * (snow pattern relative to station) * (precip at station) * (% of precip falling as snow in current grid cell). In addition to the 1-layer binary map specifying the snow pattern, each meteorology station must have a text file "StationName.snowpattern" in the meteorology directory that gives the AVERAGE snow pattern of all grid cells associated with that station (e.g., within the 4 km gridMET footprint). This is analogous to the "StationName.prism" files.

The config file now requires specification of "Snow Pattern = TRUE or FALSE" in the Options section, in addition to "Snow Pattern data path = X" to use the new functionality.