Separate snowfall and rainfall patterns
This is a relative minor changing building on my previous development of the snow pattern map. Now, instead of combining the "PRISM" pattern (not necessarily PRISM data) with the snow pattern, the monthly "PRISM" pattern is used exclusively for rain, and the single snow pattern entirely determines the snowfall distribution.

Note that the "PRISM" maps are still used for both rain and snow iff Snow Pattern = FALSE in the config file.

Also note that using the snow pattern maps REQUIRES input of the .prism and .snowpattern station-average files, even if Outside = FALSE in the config file (if Snow Pattern = FALSE and Outside = FALSE, the .prism files are not used because the station PRISM value is read from the map).