Uniform meteorology interpolation option
This is a very simple change that enables a new option for the "Interpolation" key of the config file. When set to UNIFORM, the interpolation is just a simple average of all stations. While precipitation is averaged directly, other variables like temperature can still be spatially variable due to the lapse rate.

This is particularly useful when you want to match some particular precipitation pattern.  The stations/other input meteorology can be treated as a spatial average, and then the PRISM or Snow Pattern options can be used to redistribute precipitation according to some relative pattern. Previously, the relative PRISM or Snow Pattern distribution would get imprinted on top of the underlying station variability, potentially causing double-weighting of precip extremes, e.g., precip could be increased by both the multiplier map and the underlying station distances at high elevations.