Dynamic lake storage capability
I finally developed a major new capability that I've been wanting to add for a while: dynamic lake storage! DHSVM can now simulate fully coupled lake-stream-groundwater interactions based on the local water table and a power law discharge equation applied to each lake's effective storage level. This can have a substantial effect on the hydrograph in regions with large mountain lakes, as diurnal and event-scale runoff pulses are dampened/smoothed by the lake buffering effect. The area of the lake determines how much inflow is required to change the lake level, which in turn controls lake outflow, so larger lakes have a proportionally stronger effect.

When the Lake Dynamics option is TRUE, the model expects a Lake Map File and Number of Lakes entry in the terrain section of the config file. For each lake, there should be a block with the Lake Name, Outflow Channel (matching IDs in the stream network), Power Law Scale, and Power Law Exponent. The lake map file should be binary (char) with integers identifying each unique lake (and zero elsewhere). During routing, all stream segments (except for lake outlets) that intersect a lake will empty into the lake, and this amount of water will be uniformly distributed across the surface of the lake grid cells (IExcess). Based on the depth of IExcess and the power law for each lake, the lake's outflow is subtracted from the grid cells and flows back into the user-specified outflow channel.